--- Madson/src/Components/CardPokedex.css
@@ -1,89 +1,0 @@
-.card {
-    display: flex;
-    flex-direction: column;
-    width: 250px;
-    height: 360px;
-    background-color: rgb(247, 223, 223);
-    padding: 15px;
-    border-radius: 15px;
-    background-image: url(../Img/Card2.jpg);
-    background-repeat: no-repeat;
-    background-size: 100%;
-    align-items: center;
-}
-
-.imgContainer {
-    background-image: url(../Img/Background.jpg);
-    background-repeat: no-repeat;
-    background-size: 130%;
-    background-position: bottom;
-    width: 230px;
-    height: 152px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    overflow: hidden;
-}
-
-.img {
-    border-radius: 5px;
-    width: 80%;
-}
-
-.nomePokemon {
-    font-size: 1.2em;
-    font-family: Cambria, Cochin, Georgia, Times, 'Times New Roman', serif;
-    font-weight: bolder;
-    margin-top: 0;
-    margin-bottom: 5px;
-    margin-right: 50px;
-}
-
-.btnPokedexCont {
-    height: 100px;
-    width: 230px;
-    transition: 150ms;
-}
-
-.btnPokedexCont:hover {
-    width: 201px;
-}
-
-.btnPokedex {
-    border: none;
-    background-color: transparent;
-    font-family: Cambria, Cochin, Georgia, Times, 'Times New Roman', serif;
-    font-size: 1.5em;
-    font-weight: 600;
-    color: rgb(255, 255, 255);
-    cursor: pointer;
-    transition: 150ms;
-    height: 100px;
-}
-
-.btnPokedex:hover {
-    font-size: 1.3em;
-}
-
-.btnDetalhesCont {
-    height: 50px;
-    width: 230px;
-    transition: 150ms;
-}
-
-.btnDetalhes {
-    border: none;
-    background-color: transparent;
-    font-family: Cambria, Cochin, Georgia, Times, 'Times New Roman', serif;
-    font-size: 1.4em;
-    font-weight: 600;
-    color: white;
-    cursor: pointer;
-    margin-left: 115px;
-    margin-top: 38px;
-    transition: 150ms;
-}
-
-.btnDetalhes:hover {
-    font-size: 1.3em;
-}
