--- Madson/src/Styles/Home.css
@@ -1,11 +1,0 @@
-.cards-container {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    gap: 20px;
-    margin: 30px;
-}
-
-.card {
-    box-shadow: 5px 30px 500px rgba(0, 0, 0, 0.355);
-}
