+++ pokedex/src/Components/Card.css
@@ -10,6 +10,13 @@
     background-repeat: no-repeat;
     background-size: 100%;
     align-items: center;
+    transition: 150ms;
+    scale: 100%;
+}
+
+.card:hover {
+    transform: translate3d(3px, -16px, 50px) rotateX(-2deg) rotateZ(-2deg) scale(1.08);
+    box-shadow: -20px 30px 500px rgba(0, 0, 0, 0.355);
 }
 
 .imgContainer {
@@ -31,38 +38,29 @@
 }
 
 .nomePokemon {
-    font-size: 1.2em;
-    font-family: Cambria, Cochin, Georgia, Times, 'Times New Roman', serif;
+    font-size: 1.1em;
+    font-family: Verdana, Geneva, Tahoma, sans-serif;
     font-weight: bolder;
     margin-top: 0;
     margin-bottom: 5px;
-    margin-right: 50px;
+    margin-right: 30px;
 }
 
 .btnPokedexCont {
     height: 100px;
     width: 230px;
-    transition: 150ms;
-}
-
-.btnPokedexCont:hover {
-    width: 201px;
 }
 
 .btnPokedex {
     border: none;
     background-color: transparent;
-    font-family: Cambria, Cochin, Georgia, Times, 'Times New Roman', serif;
-    font-size: 1.5em;
+    font-family: Verdana, Geneva, Tahoma, sans-serif;
+    font-size: 1.3em;
     font-weight: 600;
     color: rgb(255, 255, 255);
     cursor: pointer;
-    transition: 150ms;
     height: 100px;
-}
-
-.btnPokedex:hover {
-    font-size: 1.3em;
+    margin-top: 10px;
 }
 
 .btnDetalhesCont {
@@ -74,16 +72,11 @@
 .btnDetalhes {
     border: none;
     background-color: transparent;
-    font-family: Cambria, Cochin, Georgia, Times, 'Times New Roman', serif;
+    font-family: Verdana, Geneva, Tahoma, sans-serif;
     font-size: 1.4em;
     font-weight: 600;
     color: white;
     cursor: pointer;
     margin-left: 115px;
     margin-top: 38px;
-    transition: 150ms;
-}
-
-.btnDetalhes:hover {
-    font-size: 1.3em;
 }
