+++ pokedex/src/Components/CardPokedex.jsx
@@ -1,7 +1,7 @@
 import axios from "axios";
 import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import "./CardPokedex.css"
+import "./Card.css"
 import { GlobalContext } from "../Global/GlobalContext";
 import { useContext } from "react";
 
@@ -10,21 +10,18 @@ const Card = (props) => {
     const navigate = useNavigate()
 
     const {setters, states} = useContext(GlobalContext);
-    const {pokedex, pokeBase, request} = states
+    const {pokedex, pokeBase} = states
     const {setPokemons, setPokedex} = setters;
 
     const [pokemon, setPokemon] = useState({})
-    const [img, setImg] = useState("")
-    const [name, setName] = useState("")
 
     const url = props.url
     const index = props.index
 
-    const formataNome = () => {
-        let nome = props.nome
-        const novoNome = nome.replace(/^./, nome[0].toUpperCase());
-        
-        setName(novoNome)
+    const formataStat = (name) => {
+        let novoNome = name.replace(/^./, name[0].toUpperCase());
+
+        return novoNome
     }
 
     const onClickRemovePokedex = (i) => {
@@ -38,22 +35,20 @@ const Card = (props) => {
     useEffect(() => {
         
         axios.get(url).then((res) => {
-            setPokemon(res.data)
-            setImg(res.data.sprites.front_default)
-            formataNome()
+            setPokemon(res.data);
         }).catch((err) => {
             console.log(err);
         })
-    },[request])
+    },[])
 
     return (
         <div className="card">
-            <p className="nomePokemon">{name}</p>
+            <p className="nomePokemon">{ pokemon.name && formataStat(pokemon.name)}</p>
             <div className="imgContainer">
-                <img src={img} className="img" alt={name}/>
+                { pokemon.sprites && <img src={pokemon.sprites.front_default} className="img" alt={ pokemon.name && formataStat(pokemon.name)}/>}
             </div>
             <div className="btnPokedexCont">
-                <button className="btnPokedex" onClick={() => {onClickRemovePokedex(index)}}>Remova {name} do seu Pokedex!</button>
+                <button className="btnPokedex" onClick={() => {onClickRemovePokedex(index)}}>Remova { pokemon.name && formataStat(pokemon.name)} do seu Pokedex!</button>
             </div>
             <div className="btnDetalhesCont">
             <button className="btnDetalhes" onClick={() => {navigate(`/Detalhes/${pokemon.id}`)}}>Detalhes</button>
